+++ bot/resource/exception/handler/generic_exception_handler.py
@@ -1,5 +1,5 @@
 from bot.resource.constants import ReplyMessage
-from bot.resource.exception_handler.exception_handler import ExceptionHandler
+from bot.config.exception.exception_handler import ExceptionHandler
 
 
 class GenericHandler(ExceptionHandler):
