+++ bot/resource/exception/handler/student_already_exists_handler.py
@@ -2,7 +2,7 @@ from bot.application.student.exceptions.student_already_exist import (
     StudentAlreadyExistsException,
 )
 from bot.resource.constants import ReplyMessage
-from bot.resource.exception_handler.exception_handler import ExceptionHandler
+from bot.config.exception.exception_handler import ExceptionHandler
 
 
 class StudentAlreadyExistsExceptionHandler(ExceptionHandler):
