+++ bot/resource/exception/handler/student_already_register_handler.py
@@ -2,7 +2,7 @@ from bot.application.student.exceptions.student_already_registered_exception imp
     StudentAlreadyRegisteredException,
 )
 from bot.resource.constants import ReplyMessage
-from bot.resource.exception_handler.exception_handler import ExceptionHandler
+from bot.config.exception.exception_handler import ExceptionHandler
 
 
 class NotFoundExceptionMapper(ExceptionHandler):
