+++ bot/resource/exception/handler/student_not_found_handler.py
@@ -1,4 +1,4 @@
-from bot.resource.exception_handler.exception_handler import ExceptionHandler
+from bot.config.exception.exception_handler import ExceptionHandler
 from bot.infra.student.exception.student_not_found_exception import (
     StudentNotFoundException,
 )
